Update README.md
The main objective of this assignment is to introduce you to the use of classes, objects, and methods, thus applying the related concepts presented in the lectures. You will do so by building a simple Tic-Tac-Toe game. You will first implement a class that represents the state of the game and a function that implements the logic of the game. You will then write a main function that uses these classes to run a game.

Tic-Tac-Toe

Tic-tac-toe is a simple game commonly played by children. Two players, X and O, take turns marking the spaces in a 3×3 grid or board. The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal line wins the game. Player X is always the first to place a mark. The following example shows the progression of a game won by player X.